requirements.txt
streamlit==1.41.0
requests==2.31.0
python-dotenv==1.0.0
gtts==2.5.1
Pillow==10.4.0
deep-translator==1.11.4
openai==0.28.1
